create hooks file
